refactor: hoist Compose preview stubs to named objects per module
The Android actual impls of rememberPermissionController and
rememberLocationTracker each had an inline anonymous-object preview stub
inside their LocalInspectionMode branch, which detekt's
EmptyFunctionBlock kept flagging on no-op overrides. Pulled the stubs
out to internal top-level objects in commonMain — PreviewPermissionController
and PreviewLocationTracker — so:

- the preview-mode branch is now a one-liner (return PreviewX),
- iOS can opt into the same stub later without re-deriving it,
- the no-op overrides are documented + named instead of buried in an
  anonymous body, so future detekt complaints are addressed at the
  source rather than per-call-site.

Both stubs stay internal — production never sees them.
