+++ core/location/data/src/androidMain/kotlin/xyz/ksharma/krail/core/location/data/ComposeIntegration.android.kt
@@ -6,9 +6,6 @@ import androidx.compose.runtime.DisposableEffect
 import androidx.compose.runtime.remember
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.platform.LocalInspectionMode
-import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.emptyFlow
-import xyz.ksharma.krail.core.location.Location
 
 /**
  * Android implementation of [rememberLocationTracker].
@@ -19,19 +16,7 @@ import xyz.ksharma.krail.core.location.Location
 @Composable
 actual fun rememberLocationTracker(): LocationTracker {
     if (LocalInspectionMode.current) {
-        return remember {
-            object : LocationTracker {
-                override suspend fun getCurrentLocation(timeoutMs: Long): Location =
-                    Location(latitude = 0.0, longitude = 0.0, timestamp = 0L)
-
-                override fun startTracking(config: xyz.ksharma.krail.core.location.LocationConfig): Flow<Location> =
-                    emptyFlow()
-
-                override fun stopTracking() = Unit
-
-                override suspend fun isLocationEnabled(): Boolean = true
-            }
-        }
+        return PreviewLocationTracker
     }
 
     val context = LocalContext.current
